feat(tests): add sitemap link validation tests and update workflow
